+++ RELEASE_NOTES.md
@@ -1,6 +1,15 @@
 # kb_mash release notes
 =========================================
 
+0.0.3
+-----
+* added support for genome object as input
+* added the mash icon
+
+0.0.2
+-----
+* uses kbase sketch service
+
 0.0.0
 -----
 * Module created by kb-sdk init